chore: remove legacy sequencer command
refactor: move start command into rollapp flow
